requirements.txt
fastapi==0.83.0
starlette==0.19.1
pydantic==1.9.2
uvicorn==0.16.0
requests==2.27.1
Pillow==8.4.0
